급식 자료실(MealPolicy) 파일 부분

+++ schoolMeal/src/main/java/com/example/schoolMeal/domain/repository/mealResource/MealPolicyRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.example.schoolMeal.domain.entity.mealResource.MealPolicy;
 
 @Repository
-public interface MealPolicyRepository extends JpaRepository<MealPolicy, Integer> {
+public interface MealPolicyRepository extends JpaRepository<MealPolicy, Long> {
 }
